* Add rpc sevice to pow

* Fix mined

* Refactor the rpc api

* Fix proof missing.

* Refactr miner server

* Add miner client waker && refactor the miner client

* Refactor the miner

* Add tests

* Add more tests

* Replace oneshot channel to async_std channel

* Make miner as a mod

* Implement default for miner

* Update lib.rs to add mine_client

* Fix miner rpc server droped.

* Make miner rpc server configurable

* verify solve succ.

* cargo fmt

+++ consensus/src/pow/payload_ext.rs
@@ -9,7 +9,7 @@ use std::fmt;
 pub struct BlockPayloadExt {
     pub txns: Vec<SignedTransaction>,
     pub nonce: u64,
-    pub solve: Vec<u32>,
+    pub solve: Vec<u8>,
 }
 
 impl BlockPayloadExt {
